refactor: created VectorColorClassification component
